requirements.txt
python-magic==0.4.10
Pillow==3.2.0
cropresize2==0.1.9
short-url==1.2.1
Flask-Mako==0.4
Flask-SQLAlchemy==2.3.0